% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{contextmenu}
\alias{contextmenu}
\title{Contextmenu}
\usage{
contextmenu(x, y, ...)
}
\arguments{
\item{x, y}{\code{\link[base:integer]{Integer}}. Contextmenus upper left corner
is displayed at (\code{x} px, \code{y} px).}

\item{...}{UI elements to include in the contextmenu, typically \code{contextmenu_item}
and \code{contextmenu_hr}.}
}
\description{
This creates the UI representing a contextmenu at absolute position
(\code{x}, \code{y}). Contextmenus are typically displayed after the user
clicked UI using the right mouse button. Contextmenus consist of
\code{\link{contextmenu_item}} and \code{\link{contextmenu_hr}}.
}
