% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{contextmenu_hr}
\alias{contextmenu_hr}
\title{Contextmenu Horizontal Rule}
\usage{
contextmenu_hr()
}
\description{
This is just a wrapper for \code{\link[htmltools:builder]{htmltools::hr}}. Use
horizontal rules to structure a \code{\link{contextmenu}}.
}
