% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_callback.R
\name{datatable_callback}
\alias{datatable_callback}
\title{Callback that adds shiny inputs for specific events}
\usage{
datatable_callback(inputId)
}
\arguments{
\item{inputId}{Typically the \code{outputId} used for displaying the corresponding
\code{\link[DT]{datatable}}.}
}
\description{
The output of this function may be used as the \code{callback} argument to
\code{\link[DT]{datatable}}. Event listeners for \code{dblclick}
and \code{contextmenu} are added to the datatable. These event listeners
set a shiny input, so that you can listen to these events on the R-side
with \code{\link[shiny]{observeEvent}}. See 'Details' for the properties of
the added inputs.
}
\details{
The callback adds the following inputs: \code{input$<inputId>_row_dblclicked}
and \code{input$<inputId>_row_contextmenued}, where \code{<inputId>} is
replaced by the parameter \code{inputId} passed to this function. These inputs
have the following properties:
\describe{
  \item{\code{<inputId>_row_dblclicked}}{
    \tabular{ll}{
      \code{row} \tab The index of the dblclicked row. \cr
      \code{id} \tab The value of the first column of the dblclicked row. It is
      convenient to create unique ids for every row in the first column of a
      table passed to \code{\link[DT]{datatable}} and hiding this column via
      the \code{options} argument. \cr
      \code{data} \tab The data of the dblclicked row. \cr
      \code{rnd} \tab A random number. \cr
    }
  }
  \item{\code{<inputId>_row_contextmenued}}{
    \tabular{ll}{
      \code{row} \tab The index of the dblclicked row. \cr
      \code{id} \tab The value of the first column of the dblclicked row. It is
      convenient to create unique ids for every row in the first column of a
      table passed to \code{\link[DT]{datatable}} and hiding this column via
      the \code{options} argument. \cr
      \code{data} \tab The data of the dblclicked row. \cr
      \code{mouse} \tab A list with elements named \code{x} and \code{y}
      holding the position of the mouse in px, when the datatable was
      contextmenued. \cr
      \code{rnd} \tab A random number. \cr
    }
  }
}
}
