% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_body.R
\name{explorer_body}
\alias{explorer_body}
\alias{explorer_body_ui}
\title{Explorer Body}
\usage{
explorer_body_ui(id)

explorer_body(input, output, session, .values, .parent, .children_r,
  .root_node_r, .explorer_rvs, .addable_explorer_classes_r,
  .visible_explorer_classes_r)
}
\arguments{
\item{id}{The module's id.}

\item{input, output, session}{Called by \code{\link[shiny]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{\link{Node}} object.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}.}

\item{.explorer_rvs}{An object of class \code{\link[shiny]{reactiveValues}}.
See 'Details' for list of needed elements.}

\item{.addable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are addable to the explorer.}

\item{.visible_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are displayed to the user. Group nodes are always
displayed.}
}
\description{
You usually don't want to call this module explicity.
}
\details{
Shiny module called by \code{\link{explorer}} representing a
\code{\link[DT]{datatable}} displaying the elements of the node tree with
root node \code{.root_node_r()}.


\code{.explorer_rvs} must provide the following elements:
\tabular{ll}{
  \code{current_node} \tab \code{\link{ExplorerNode}}, which is currently
  selected in \code{\link{explorer}}. \cr
  \code{contextmenued_node} \tab \code{\link{ExplorerNode}}, which has been
  last contextmenued. \cr
  \code{module_ids} \tab \code{\link[base:list]{List}} with one entry per
  unique explorer class.
}
}
