% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplorerClass.R
\docType{data}
\name{ExplorerClass}
\alias{ExplorerClass}
\title{Explorer Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExplorerClass
}
\description{
R6Class, whose instantiations represent an  class inside the
\code{\link{explorer}} module. The following list describes the arguments
of \code{ExplorerClass$new}.
}
\section{Server function}{

The following (and only these) return list elements are handled by
\code{\link{explorer}} and \code{\link{explorer_body}}:
\tabular{ll}{
  \code{selectable_classes_r} \tab \code{\link[shiny:reactive]{Reactive}}
  returning a character vector containing the \code{id}s of \code{explorer_classes}
  who can be added to the \code{explorer}, if the current \code{explorer_class}
  has \code{id = id}. Returning \code{\link[base:NULL]{NULL}} results in no
  possible selection. Returning of \code{""} results in possible selection of
  all \code{explorer_classes}. If the server function does not return this
  reactive, all \code{explorer_classes} are selectable. \cr
  \code{icon_r} \tab \code{\link[shiny:reactive]{Reactive}} returning a
  \code{\link[shiny:icon]{icon}}, which is displayed in the \code{explorer_body}.
}
}

\keyword{datasets}
