% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplorerTree.R
\name{ExplorerTree}
\alias{ExplorerTree}
\title{Explorer Tree}
\description{
Explorer Tree
}
\section{Methods}{

\describe{
  \item{\code{new(root_id = NULL, root_object = Object$new("/"))}}{Initialize a new explorer tree object.
    \tabular{ll}{
      \code{root_id} \tab Identifier of the root node. If \code{\link[base]{NULL}},
      this identifier is created internally.
    }
    Return: Object of class \code{ExplorerTree}.
  }
  \item{\code{get_node(id)}}{Get the node of the explorer tree with \code{id == id}.
  }
  \item{\code{get_node_ids()}}{Get the ids of all nodes of the explorer tree as a
    \code{\link[base]{character}} vector.
  }
  \item{\code{get_root_node()}}{Get the root node of the explorer tree, which is equivalent
    to calling \code{this$get_node(root_id)}, where \code{root_id} is the id, which was
    passed to \code{this$new}.
  }
}
}

