% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_classes.R
\name{group_explorer_class}
\alias{group_explorer_class}
\title{Group Explorer Class}
\usage{
group_explorer_class(
  id = "__group__",
  .label_list = label_group_explorer_class()
)
}
\arguments{
\item{.label_list}{A \code{\link[base]{list}} created with
\code{\link{label_group_explorer_class}} containing labels for all buttons
used inside the group_explorer_class.}
}
\description{
Group Explorer Class
}
