% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_group_node}
\alias{label_group_node}
\title{Create Labels for Group Node Explorer Class}
\usage{
label_group_node(confirm_rename = "Rename", open = "Open",
  rename = "Rename")
}
\arguments{
\item{confirm_rename, open, rename}{\code{\link[base:character]{Character}}
string that will appear on the corresponding labels.}
}
\description{
Create Labels for Group Node Explorer Class
}
