% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_rename_contextmenu_item}
\alias{label_rename_contextmenu_item}
\title{Create Labels for the Rename Contextmenu Item}
\usage{
label_rename_contextmenu_item(confirm_rename = "Confirm",
  start_rename = "Rename", title_rename = "Rename")
}
\arguments{
\item{confirm_rename, start_rename, title_rename}{\code{\link[base:character]{Character}}
strings that will appear on the corresponding labels.}
}
\description{
Create Labels for the Rename Contextmenu Item
}
