% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectStorage.R
\name{ObjectStorage}
\alias{ObjectStorage}
\title{ObjectStorage}
\description{
R6Class for storing similar R6 objects. These objects all need to implement
the method get_id() and get_name(). The id of an object is
unique in a storage whereas multiple objects can share the same name.
}
\section{Usage}{

\preformatted{storage = ObjectStorage$new(allowed_classes = NULL)

storage$add_object(object)

storage$get_object(id)
}
}

