% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_input.R
\name{positional_input}
\alias{positional_input}
\title{Positional Input}
\usage{
positional_input(inputId, label, icon = NULL)
}
\description{
A Shiny input behaving similarly to \code{\link[shiny]{actionButton}}. In
addition to \code{input$<inputId>} you may use \code{input$<inputId>_position},
which returns a list
}
