% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_contextmenu_item.R
\name{rename_contextmenu_item}
\alias{rename_contextmenu_item}
\alias{rename_contextmenu_item_ui}
\title{Contextmenu item for renaming}
\usage{
rename_contextmenu_item_ui(id,
  .label_list = label_rename_contextmenu_item())

rename_contextmenu_item(input, output, session, .values, .explorer_rvs,
  .label_list = label_rename_contextmenu_item())
}
\arguments{
\item{id}{Module id.}

\item{.label_list}{A list created with \code{\link{label_rename_contextmenu_item}}
containing labels for all UI elements used inside this module.}

\item{input, output, session}{Calley by \code{\link[shiny]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.explorer_rvs}{\code{\link[shiny:reactiveValues]{ReactiveValues}}.}
}
\description{
Clicking on this contextmenu item starts the renaming process for the
associated object of the contextmenued object of class \code{\link{ExplorerNode}}.
}
