% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run shinyExplorer examples}
\usage{
run_example(example)
}
\arguments{
\item{example}{The name of the example app to launch.}
}
\description{
Launch a \code{shinyExplorer} example Shiny app thta shows how to use
\code{shinyExplorer} in an app.
}
\details{
Call this function without arguments to see a list of available example apps.
}
