% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyExplorer.R
\name{shinyExplorer}
\alias{shinyExplorer}
\title{Run shinyMlr}
\usage{
shinyExplorer(...)
}
\arguments{
\item{...}{[\code{any}]\cr
Additional arguments passed to shiny's
\code{runApp()} function.}
}
\description{
Run a local instance of shinyMlr.
}
\examples{
\dontrun{
  shinyExplorer()
}
}
