% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{show_contextmenu}
\alias{show_contextmenu}
\title{Show Contextmenu}
\usage{
show_contextmenu(contextmenu, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{contextmenu}{A \code{\link{contextmenu}}.}

\item{session}{The \code{session} object passed fo function given to
\code{shinyServer}.}
}
\description{
This causes a contextmenu to be displayed in the client browser, and is
typically used with \code{\link{contextmenu}}. \code{\link{use_contextmenu}}
must be called somewhere in your UI, prior to calling this function.
}
