% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextmenu.R
\name{use_contextmenu}
\alias{use_contextmenu}
\title{Use Contextmenu}
\usage{
use_contextmenu()
}
\description{
Call this function in your UI to enable the functionality of
\code{\link{show_contextmenu}}.
}
