% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\title{Explorer}
\usage{
explorer(input, output, session, .values, .parent, .root_node_r,
  .group_nodes_addable = TRUE,
  .addable_explorer_classes_r = shiny::reactive(NULL),
  .visible_explorer_classes_r = shiny::reactive(NULL),
  .display_header = TRUE)
}
\arguments{
\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.parent}{The parent \code{Node} object.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}.}

\item{.group_nodes_addable}{\code{\link[base:logical]{Logical}} indicating
whether group nodes are addable or not.}

\item{.addable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are addable to the explorer.}

\item{.visible_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are displayed to the user. Group nodes are always
displayed.}

\item{.display_header}{If \code{\link[base:logical]{TRUE}}, the navigation
header is displayed, otherwise it is not.}
}
\description{
Shiny module representing an explorer usable for selecting elements of different
kinds and invoking actions on them.
}
