% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer.R
\name{explorer}
\alias{explorer}
\alias{explorer_ui}
\title{Explorer}
\usage{
explorer_ui(id)

explorer(input, output, session, .values, .root_node_r, .explorer_classes,
  .group_nodes_addable = TRUE,
  .addable_explorer_classes_r = shiny::reactive(NULL),
  .visible_explorer_classes_r = shiny::reactive(NULL),
  .display_header = TRUE, .label_list = label_explorer())
}
\arguments{
\item{id}{The module's id.}

\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}.}

\item{.explorer_classes}{A \code{\link[base]{list}} of objects of class
\code{\link{ExplorerClass}}.}

\item{.group_nodes_addable}{\code{\link[base:logical]{Logical}} indicating
whether group nodes are addable or not.}

\item{.addable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are addable to the explorer.}

\item{.visible_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are displayed to the user. Group nodes are always
displayed.}

\item{.display_header}{If \code{\link[base:logical]{TRUE}}, the navigation
header is displayed, otherwise it is not.}

\item{.label_list}{A \code{\link[base]{list}} created with \code{\link{label_explorer}}
containing labels for all buttons used inside the explorer module.}
}
\description{
Shiny module representing an explorer usable for selecting elements of different
kinds and invoking actions on them.
}
