% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_body.R
\name{explorer_body}
\alias{explorer_body}
\alias{explorer_body_ui}
\title{Explorer Body}
\usage{
explorer_body_ui(id)

explorer_body(input, output, session, .values, .parent, .children_r,
  .root_node_r, .explorer_rvs, .addable_explorer_classes_r,
  .visible_explorer_classes_r)
}
\description{
Explorer Body
}
