% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_body.R
\name{explorer_body}
\alias{explorer_body}
\alias{explorer_body_ui}
\title{Explorer Body}
\usage{
explorer_body_ui(id)

explorer_body(input, output, session, .values, .children_r, .root_node_r,
  .explorer_classes, .explorer_class_returns, .explorer_rvs,
  .addable_explorer_classes_r, .visible_explorer_classes_r, .label_list)
}
\arguments{
\item{id}{Module id.}

\item{input, output, session}{Called by \code{\link[shiny]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.children_r}{\code{\link[shiny]{reactive}} containing the child nodes of
the current node.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}. This needs not necessarily to be the
root node of an object of class \code{\link{ExplorerTree}}.}

\item{.explorer_classes}{A \code{\link[base]{list}} of objects of class
\code{\link{ExplorerClass}}.}

\item{.explorer_rvs}{\code{\link[shiny:reactiveValues]{ReactiveValues}}.}

\item{.addable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are addable to the explorer.}

\item{.visible_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are displayed to the user. Group nodes are always
displayed.}

\item{.label_list}{A \code{\link[base]{list}} created with \code{\link{label_explorer}}
containing labels for all buttons used inside the explorer module.}
}
\description{
You usually don't want to call this module explicitly.
}
\details{
#' \code{.explorer_rvs} must provide the following elements:
\tabular{ll}{
  \code{current_node} \tab \code{\link{ExplorerNode}}, which is currently
  selected in \code{\link{explorer}}. \cr
  \code{contextmenued_node} \tab \code{\link{ExplorerNode}}, which has been
  last contextmenued. \cr
  \code{module_ids} \tab \code{\link[base:list]{List}} with one entry per
  unique explorer class.
}
}
