% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_selector.R
\name{explorer_selector}
\alias{explorer_selector}
\alias{explorer_selector_ui}
\title{Explorer Selector}
\usage{
explorer_selector_ui(id)

explorer_selector(input, output, session, .values, .root_node_r,
  .explorer_classes, .group_nodes_selectable = FALSE,
  .selectable_explorer_classes_r = shiny::reactive(character()),
  .addable_explorer_classes_r = .selectable_explorer_classes_r,
  .visible_explorer_classes_r = .selectable_explorer_classes_r,
  .label_list = label_explorer_selector())
}
\arguments{
\item{id}{Module id.}

\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}. This needs not necessarily to be the
root node of an object of class \code{\link{ExplorerTree}}.}

\item{.explorer_classes}{A \code{\link[base]{list}} of objects of class
\code{\link{ExplorerClass}}.}

\item{.group_nodes_selectable}{\code{\link[base:logical]{Logical}} indicating
whether group nodes are selectable or not.}

\item{.selectable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of selectable explorer_classes. If \code{character()}, only group nodes are
selectible (if \code{.group_nodes_selectable = TRUE})}

\item{.addable_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are addable to the explorer.}

\item{.visible_explorer_classes_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the ids
of explorer classes that are displayed to the user. Group nodes are always
displayed.}

\item{.label_list}{A \code{\link[base]{list}} created with \code{\link{label_explorer}}
containing labels for all buttons used inside the explorer module.}
}
\description{
Use the \code{\link{explorer}} module to select a node.
}
