% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorer_selector.R
\name{explorer_selector}
\alias{explorer_selector}
\alias{explorer_selector_ui}
\title{Explorer Selector}
\usage{
explorer_selector_ui(id)

explorer_selector(
  input,
  output,
  session,
  .values,
  .root_node_r,
  .explorer_classes = list(),
  selectable_r = shiny::reactive(character()),
  addable_r = shiny::reactive(character()),
  visible_r = shiny::reactive(character()),
  .label_list = label_explorer_selector(),
  ui = c("default", "minimal"),
  mode = c("modal", "in_place"),
  action_button_fun = shiny::actionButton
)
}
\arguments{
\item{id}{Module id.}

\item{input, output, session}{Called by \code{\link[shiny:callModule]{callModule}}.}

\item{.values}{The \code{.values} list.}

\item{.root_node_r}{A \code{\link[shiny:reactive]{reactive}} returning an
object of class \code{\link{ExplorerNode}}. This needs not necessarily to be the
root node of an object of class \code{\link{ExplorerTree}}.}

\item{.explorer_classes}{A \code{\link[base]{list}} of objects of class
\code{\link{ExplorerClass}}.}

\item{selectable_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing labels
of selectable explorer_classes.}

\item{addable_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the labels
of explorer classes that are addable to the explorer.}

\item{visible_r}{A \code{\link[shiny:reactive]{reactive}}
returning a \code{\link[base:character]{character}} vector containing the labels
of explorer classes that are displayed to the user.}

\item{.label_list}{A \code{\link[base]{list}} created with \code{\link{label_explorer}}
containing labels for all buttons used inside the explorer module.}

\item{ui}{If \code{"default"}, the UI consists of an actionButton for
selecting an element and an \code{\link{explorer}}, in which the selected
element is shown. If \code{"minimal"}, only an actionButton showing the
name of the selected element as label is shown.}

\item{mode}{If \code{"modal"} a modal is opened for selecting an element. If
\code{"in_place"} an explorer is opened in place for selecting an element.}

\item{action_button_fun}{A \code{\link[base]{function}} returning an HTML button,
which is connected with shiny.}
}
\value{
The \code{explorer_selector} module returns a list containing the following
reactives:
\tabular{ll}{
  \code{selected_node_r} \tab An object of class \code{\link{ExplorerNode}},
  which has been selected by the user. If no node is selected, the computation
  is stopped by \code{\link[shiny]{req}}. \cr
  \code{selected_node_or_null_r} \tab An object of class \code{\link{ExplorerNode}},
  which has been selected by the user or \code{\link[base]{NULL}}, if no node
  is selected.
}
}
\description{
Use the \code{\link{explorer}} module to select a node.
}
