% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplorerClass.R
\docType{data}
\name{ExplorerClass}
\alias{ExplorerClass}
\title{Explorer Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExplorerClass
}
\arguments{
\item{id}{A \code{\link[base:character]{character}}. See 'Details' for use of
the id.}

\item{ui}{A list of functions with argument \code{id} returning a
\code{\link{contextmenu_item}}, which is displayed if the current node of
the \code{\link{explorer}} has a child node, whose object is from
\code{id = id} inside of the \code{\link{contextmenu}}. Clicking on it will
make this child node the current node.}

\item{server}{Module server function. See 'Server function' for return list
elements that get handled by \code{\link{explorer}} and \code{\link{explorer_body}}.}
}
\description{
R6Class, whose instantiations represent an  class inside the
\code{\link{explorer}} module. The following list describes the arguments
of \code{ExplorerClass$new}.
}
\section{Server function}{

The following (and only these) return list elements are handled by
\code{\link{explorer}} and \code{\link{explorer_body}}:
\tabular{ll}{
  \code{selectable_classes_r} \tab \code{\link[shiny:reactive]{Reactive}}
  returning a character vector containing the \code{id}s of \code{explorer_classes}
  who can be added to the \code{explorer}, if the current \code{explorer_class}
  has \code{id = id}. Returning \code{\link[base:NULL]{NULL}} results in no
  possible selection. Returning of \code{""} results in possible selection of
  all \code{explorer_classes}. If the server function does not return this
  reactive, all \code{explorer_classes} are selectable. \cr
  \code{icon_r} \tab \code{\link[shiny:reactive]{Reactive}} returning a
  \code{\link[shiny:icon]{icon}}, which is displayed in the \code{explorer_body}.
}
}

\keyword{datasets}
