% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplorerTree.R
\name{ExplorerTree}
\alias{ExplorerTree}
\title{Explorer Tree}
\description{
Explorer Tree
}
\section{Methods}{

\describe{
  \item{\code{new(id, root_id = NULL)}}{Initialize a new explorer tree object.
    \tabular{ll}{
      \code{id} \tab Unique \code{\link[base]{character}} string. \cr
      \code{root_id} \tab Either unique \code{\link[base]{character}} string within this
      explorer tree or \code{\link[base]{NULL}} for the root node of the explorer tree.
      If \code{NULL} an id is chosen internally. You don't need to worry about the uniqueness
      of \code{root_id} since you are just creating the first node of the tree.
    }
    Return: Object of class \code{ExplorerTree}.
  }
  \item{\code{get_id()}}{Get the explorer tree's id.
  }
  \item{\code{get_node(id)}}{Get the node of the explorer tree with \code{id == id}.
  }
  \item{\code{get_node_ids()}}{Get the ids of all nodes of the explorer tree as a
    \code{\link[base]{character}} vector.
  }
  \item{\code{get_root_node()}}{Get the root node of the explorer tree, which is equivalent
    to calling \code{this$get_node(root_id)}, where \code{root_id} is the id, which was
    passed to \code{this$new}.
  }
}
}

