% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label_explorer}
\alias{label_explorer}
\title{Create Labels for Explorer}
\usage{
label_explorer(
  add_group = "New group",
  ask_delete = "Are you sure you want to delete the selected item?",
  confirm_delete = "Confirm",
  delete_node = "Delete"
)
}
\arguments{
\item{add_group, ask_delete, confirm_delete, delete_node}{\code{\link[base:character]{Character}}
strings that will appear on the corresponding labels.}
}
\description{
Create Labels for Explorer
}
